Rename SCPAPI.py to scp_api.py per PEP8 standard
